reorganize tests with mocks into tests/with-mocks
