+++ tests/with-mocks/lib/create/with-example/with-logging/with-options/create-with-example-with-options.test.js
@@ -1,4 +1,4 @@
-const lib = require('../../lib/lib.js');
+const lib = require('../../../../../../../lib/lib.js');
 
 // special compact mocks for this test:
 const mysnap = [];
