+++ tests/with-mocks/lib/create/with-options/for-windows/create-with-options-for-windows.test.js
@@ -1,6 +1,10 @@
-const func = require('../../../lib/cli-command.js').func;
+const lib = require('../../../../../../lib/lib.js');
 
-// special compact mocks for this test:
+// special compact mocks for this test
+// with mock uuid for Windows:
+jest.mock('uuid', () => ({
+  v1: () => 'E22606E0-B47F-11E9-A3F0-07F70A25DAFB'
+}));
 const mysnap = [];
 const mockpushit = x => mysnap.push(x);
 jest.mock('fs-extra', () => ({
@@ -14,19 +18,17 @@ jest.mock('fs-extra', () => ({
   },
 }));
 
-test('create alice-bobbi module with explicit config options for Android & iOS', () => {
-  const args = ['alice-bobbi'];
-
-  const config = 'bogus';
-
+test('create alice-bobbi module using mocked lib with config options on Windows', () => {
   const options = {
-    platforms: 'android,ios',
+    platforms: ['windows'],
+    name: 'alice-bobbi',
     githubAccount: 'alicebits',
     authorName: 'Alice',
     authorEmail: 'contact@alice.me',
     license: 'ISC',
+    namespace: 'Carol',
   };
 
-  return func(args, config, options)
+  return lib(options)
     .then(() => { expect(mysnap).toMatchSnapshot(); });
 });
